prod changes

+++ application/models/gcal_model.php
@@ -98,6 +98,10 @@ class Gcal_model extends CI_Model {
   }
 
   public function delete($id) {
+    if ($id == NULL) {
+      return true;
+    }
+
     $event = $this->service->getCalendarEventEntry($id);
     return $event->delete();
   }
